requirements.txt
# Core scientific and ML stack
numpy==1.23.5
pandas==1.5.3
matplotlib==3.6.3
scikit-learn==1.2.2

# Deep learning
tensorflow==2.12.0
keras==2.12.0

# Computer vision
opencv-python==4.7.0.72
Pillow==9.5.0

# Web backend
Flask==2.0.3
Flask-RESTful==0.3.9
Werkzeug==2.0.3
Jinja2==3.1.2
click==8.1.3

# Utilities
requests==2.31.0

# Selenium for scraping (if you use it)
selenium==4.10.0